requirements.txt
fastapi
uvicorn
httpx
openai-whisper
pydub